Refactored test_api_put
