+++ tests/test_api_auth.py
@@ -2,10 +2,10 @@
 from base64 import b64encode
 from json import dumps
 
-from .test_utils import build_token_auth_header
-from .test_utils import HOST_URL
-from .test_utils import set_environment
 from app.auth.identity import Identity
+from tests.utils import set_environment
+from tests.utils.api_utils import build_token_auth_header
+from tests.utils.api_utils import HOST_URL
 
 
 def valid_identity():
