+++ tests/test_mq_service.py
@@ -14,10 +14,10 @@ from app.queue.queue import _validate_json_object_for_utf8
 from app.queue.queue import event_loop
 from app.queue.queue import handle_message
 from lib.host_repository import AddHostResult
-from tests.test_utils import assert_mq_host_data
-from tests.test_utils import generate_uuid
-from tests.test_utils import minimal_host
-from tests.test_utils import valid_system_profile
+from tests.utils import generate_uuid
+from tests.utils import valid_system_profile
+from tests.utils.api_utils import minimal_host
+from tests.utils.mq_utils import assert_mq_host_data
 
 
 def test_event_loop_exception_handling(mocker, flask_app):
