+++ tests/test_tagging.py
@@ -3,13 +3,13 @@ from unittest import main
 
 from sqlalchemy import null
 
-from .test_api_utils import PaginationBaseTestCase
-from .test_api_utils import PreCreatedHostsBaseTestCase
-from .test_utils import generate_uuid
-from .test_utils import HOST_URL
-from .test_utils import inject_qs
 from app import db
 from app.models import Host
+from tests.test_api_utils import PaginationBaseTestCase
+from tests.test_api_utils import PreCreatedHostsBaseTestCase
+from tests.utils import generate_uuid
+from tests.utils.api_utils import HOST_URL
+from tests.utils.api_utils import inject_qs
 
 
 class TagsPreCreatedHostsBaseTestCase(PreCreatedHostsBaseTestCase):
