+++ tests/test_unit.py
@@ -46,7 +46,7 @@ from app.serialization import serialize_canonical_facts
 from app.serialization import serialize_host
 from app.serialization import serialize_host_system_profile
 from app.utils import Tag
-from tests.test_utils import set_environment
+from tests.utils import set_environment
 
 
 class ApiOperationTestCase(TestCase):
